GMT_DOS-SYSTEMS_M2 ASSEMBLY -> ASMS  * rename `assembly` module as `asms`
